Refactored maps

+++ ECA1D/Rule110Composite.h
@@ -4,15 +4,15 @@
 #include <map>
 
 #include "Rule110Glider.h"
-#include "Rule110CompositeMaps.h"
+#include "Rule110GliderSet.h"
 
 using namespace std;
 
 class Rule110Composite {
 public:
 	static string GetComposite(string _token);
-	static string GetGlider(int _multiple, string _gliderId, char _gliderPhase, int _gliderPhaseNum, int _etherPhase);
-	static string GetGliderSet(string _gliderSetId, char _gliderPhase, int _gliderPhaseNum, int _etherPhase);
+	static string GetGlider(int _multiple, string _gliderId, char _gliderPhase, size_t _gliderPhaseNum, size_t _etherPhase);
+	static string GetGliderSet(string _gliderSetId, char _gliderPhase, size_t _gliderPhaseNum, size_t _etherPhase);
 
 	static bool IsComposite(string& _s);
 
@@ -20,5 +20,5 @@ public:
 
 private:
 	static void processGlider(int& multiple, string& gliderId, string& _s);
-	static void processPhase(char& gliderPhase, int& gliderPhaseNum, int& etherPhase, string _s);
+	static void processPhase(char& gliderPhase, size_t& gliderPhaseNum, size_t& etherPhase, string _s);
 };
